requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Forecasting & Machine Learning
scikit-learn==1.3.0
statsmodels==0.14.0
prophet==1.1.4
pmdarima==2.0.3

# Web Framework
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
python-multipart==0.0.6

# Dashboard
streamlit==1.27.0
streamlit-aggrid==0.3.4.post3

# Visualization
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2

# Data Validation
great-expectations==0.17.15

# Optimization
cvxpy==1.3.2
scipy==1.11.1

# File Handling
openpyxl==3.1.2
xlsxwriter==3.1.2

# Utilities
requests==2.31.0
aiohttp==3.8.5
tqdm==4.66.1
joblib==1.3.2

# Database (Optional)
sqlalchemy==2.0.20
psycopg2-binary==2.9.7

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code Quality
black==23.9.1
flake8==6.1.0
mypy==1.5.1

# Documentation
mkdocs==1.5.2
mkdocs-material==9.4.0
